+++ app/Models/UserStore.php
@@ -7,6 +7,7 @@ use Illuminate\Database\Eloquent\Model;
 use App\Models\Store;
 use App\Models\User;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 
 class UserStore extends Model
 {
@@ -32,5 +33,15 @@ class UserStore extends Model
     public function store(): BelongsTo
     {
         return $this->belongsTo(Store::class);
+    }
+
+    /**
+     * Retrieve the stores model
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\BelongsToMany
+     */
+    public function retrieveStores()
+    {
+        return $this->belongsToMany(Store::class);
     }
 }